First "working"/released version
Can add/update programs and plants via the web GUI, the system will
execute those programs.

There are still many things to refactor and add, I won't list them here

+++ WateringSystemEsp/src/httpModule/dispatcher/HandlerExecutionChain.h
@@ -77,7 +77,7 @@ public:
 		Serial.println("Calling the HandlerInterceptors post handle");
 #endif
 		//HandlerInterceptors post handle
-		for (int i = 0; i < _handlerInterceptors.size(); i++) {
+		for (size_t i = 0; i < _handlerInterceptors.size(); i++) {
 			_handlerInterceptors[i]->postHandle(req, *response, *_controller,*gardenAcceptable);
 		}
